src/api_server/requirements.txt
uvicorn
fastapi
pydantic_settings
slowapi
requests
python-multipart
pillow